requirements.txt
openai
web3
python-dotenv
py-solc-x